requirements.txt
requests==2.21.0
beautifulsoup4==4.6.3
PyYAML==3.13
